ndk

Module event

Bindings for AInputEvent, AKeyEvent and AMotionEvent

Most of these operations directly wrap functions in the NDK.

See also the Java docs for android.view.InputEvent, android.view.MotionEvent, and android.view.KeyEvent.

Structs

ButtonState

A bitfield representing the state of buttons during a motion event.

EdgeFlags

A bitfield representing which edges were touched by a motion event.

HistoricalMotionEvent

Represents a view into a past moment of a motion event

HistoricalMotionEventsIter

An iterator over all the historical moments in a MotionEvent.

HistoricalPointer

A view into a pointer at a historical moment

HistoricalPointersIter

An iterator over the pointers in a historical motion event

KeyEvent

A key event

KeyEventFlags

Flags associated with KeyEvent.

MetaState

A bitfield representing the state of modifier keys during an event.

MotionEvent

A motion event

MotionEventFlags

Flags associated with this MotionEvent.

Pointer

A view into the data of a specific pointer in a motion event.

PointersIter

An iterator over the pointers in a MotionEvent.

SourceClass

Flags representing the class of an InputEvent Source.

Enums

Axis

An axis of a motion event.

InputEvent

A native AInputEvent *

KeyAction

Key actions. See the NDK docs

Keycode

Key codes.

MotionAction

A motion action.

Source

An enum representing the source of an InputEvent.

ToolType

The tool type of a pointer.