ndk

Module looper

Bindings for ALooper

In Android, ALoopers are inherently thread-local. Due to this, there are two different ALooper interfaces exposed in this module:

ThreadLooper, which has methods for the operations performable with a looper in one’s own thread; and

ForeignLooper, which has methods for the operations performable with any thread’s looper.

Structs

FdEvent

Flags for file descriptor events that a looper can monitor.

ForeignLooper

A native ALooper *, not necessarily allocated with the current thread.

LooperError

ThreadLooper

A thread-local native ALooper *. This promises that there is a looper associated with the current thread.

Enums

Poll

The poll result from a ThreadLooper.