ndk::looper

Enum Poll

pub enum Poll<'fd> {
    Wake,
    Callback,
    Timeout,
    Event {
        ident: i32,
        fd: BorrowedFd<'fd>,
        events: FdEvent,
        data: *mut c_void,
    },
}

The poll result from a ThreadLooper.

Variants

Wake

This looper was woken using ForeignLooper::wake()

Callback

For ThreadLooper::poll_once*(), an event was received and processed using a callback.

Timeout

For ThreadLooper::poll_*_timeout(), the requested timeout was reached before any events.

Event

An event was received

Fields

ident: i32

fd: BorrowedFd<'fd>

Safety

The caller should guarantee that this file descriptor remains open after it was added via ForeignLooper::add_fd() or ForeignLooper::add_fd_with_callback().

events: FdEvent

data: *mut c_void