ndk

Module native_window

Bindings for ANativeWindow

Structs

NativeWindow

It is the C counterpart of the android.view.Surface object in Java, and can be converted both ways. Depending on the consumer, images submitted to NativeWindow can be shown on the display or sent to other consumers, such as video encoders.

NativeWindowBufferLockGuard

Lock holding the next drawing surface for writing. It is unlocked and posted on drop().

Type Aliases

Rect