ndk::native_window

Struct NativeWindowBufferLockGuard

pub struct NativeWindowBufferLockGuard<'a> { /* private fields */ }

Lock holding the next drawing surface for writing. It is unlocked and posted on drop().

Implementations

impl<'a> NativeWindowBufferLockGuard<'a>

pub fn width(&self) -> usize

The number of pixels that are shown horizontally.

pub fn height(&self) -> usize

pub fn stride(&self) -> usize

The number of pixels that a line in the buffer takes in memory.

This may be >= width.

pub fn format(&self) -> HardwareBufferFormat

The format of the buffer. One of HardwareBufferFormat.

pub fn bits(&mut self) -> *mut c_void

The actual bits.

This points to a memory segment of stride() * height() * HardwareBufferFormat::bytes_per_pixel() bytes.

Only width() pixels are visible for each stride() line of pixels in the buffer.

See bytes() for safe access to these bytes.

pub fn bytes(&mut self) -> Option<&mut [MaybeUninit<u8>]>

Safe write access to likely uninitialized pixel buffer data.

Returns None when there is no HardwareBufferFormat::bytes_per_pixel() size available for this format().

The returned slice consists of stride() * height() * HardwareBufferFormat::bytes_per_pixel() bytes.

Only width() pixels are visible for each stride() line of pixels in the buffer.

pub fn lines(&mut self) -> Option<impl Iterator<Item = &mut [MaybeUninit<u8>]>>

Returns a slice of bytes for each line of visible pixels in the buffer, ignoring any padding pixels incurred by the stride.

See bits() and bytes() for contiguous access to the underlying buffer.